move descriptions into platforms
